+++ resolwe_bio/processes/workflows/geo_import.py
@@ -1,5 +1,6 @@
 """GEO import pipeline."""
 import re
+from pathlib import Path
 
 import GEOparse
 import pandas as pd
@@ -7,6 +8,7 @@ import requests
 
 from resolwe.process import (
     BooleanField,
+    FileField,
     GroupField,
     IntegerField,
     Process,
@@ -53,6 +55,7 @@ def construct_descriptor(metadata, sample_name):
     """
 
     sample_metadata = metadata.loc[sample_name]
+    sample_metadata = sample_metadata.fillna("")
     descriptor = {"general": {}, "experiment": {}}
     # TODO Replace fixed values with calls to descriptor schema once available.
     # Also organ / tissue should be added then.
@@ -146,10 +149,21 @@ class GeoImport(Process):
     WARNING: Additional costs for storage and processing may be incurred
     if a very large data set is selected.
 
-    This runs SRA import process for each individual experiment (SRX)
-    from the selected GEO Series. In addition metadata table with sample
-    information is created and uploadad to the same collection.
-    Currently only RNA-seq datasets are supported.
+    RNA-seq and expression microarray datasets can be uploaded. For
+    RNA-Seq data sets this runs the SRA import process for each
+    experiment (SRX) from the selected RNA-Seq GEO Series.
+    If GSE contains microarray data, it downloads individual samples and
+    uploads them as microarray expression objects. Probe IDs can be
+    mapped to the Ensembl IDs if the corresponding GPL platform is
+    supported, otherwise, a custom mapping file should be provided.
+    Currently supported platforms are: GPL74, GPL201, GPL96, GPL571,
+    GPL97, GPL570, GPL91, GPL8300, GPL92, GPL93, GPL94, GPL95, GPL17586,
+    GPL5175, GPL80, GPL6244, GPL16686, GPL15207, GPL1352, GPL11068,
+    GPL26966, GPL6848, GPL14550, GPL17077, GPL16981, GPL13497, GPL6947,
+    GPL10558, GPL6883, GPL13376,GPL6884, GPL6254.
+
+    In addition metadata table with sample information is created and
+    uploaded to the same collection.
     """
 
     slug = "geo-import"
@@ -168,7 +182,7 @@ class GeoImport(Process):
         },
     }
     data_name = "{{ gse_accession }}"
-    version = "1.0.1"
+    version = "2.0.0"
     process_type = "data:geo"
     category = "Import"
     scheduling_class = SchedulingClass.BATCH
@@ -197,6 +211,31 @@ class GeoImport(Process):
             aligned = BooleanField(label="Dump only aligned sequences", default=False)
             unaligned = BooleanField(
                 label="Dump only unaligned sequences", default=False
+            )
+            mapping_file = FileField(
+                label="File with probe ID mappings",
+                description="The file should be tab-separated and contain two columns with their column names. The "
+                "first column should contain Gene IDs and the second one should contain probe names. Supported file "
+                "extensions are .tab.*, .tsv.*, .txt.*",
+                required=False,
+            )
+            source = StringField(
+                label="Gene ID source",
+                description="Gene ID source used for probe mapping is required when using a custom file.",
+                allow_custom_choice=True,
+                required=False,
+                choices=[
+                    ("AFFY", "AFFY"),
+                    ("DICTYBASE", "DICTYBASE"),
+                    ("ENSEMBL", "ENSEMBL"),
+                    ("NCBI", "NCBI"),
+                    ("UCSC", "UCSC"),
+                ],
+            )
+            build = StringField(
+                label="Genome build",
+                description="Genome build of mapping file is required when using a custom file.",
+                required=False,
             )
 
         advanced = GroupField(
@@ -280,6 +319,61 @@ class GeoImport(Process):
             sample_info.items(), columns=["Accession", "SampleName"]
         ).set_index("Accession", drop=False)
 
+    def upload_ma_gse(self, inputs, gse):
+        """Upload microarray samples from a GEO series."""
+        unmapped_data = []
+        for name, gsm in gse.gsms.items():
+            exp_path = f"{name}.tsv"
+            gsm.table.to_csv(exp_path, sep="\t", index=False)
+
+            platform_id = gsm.metadata["platform_id"][0]
+            process_inputs = {
+                "exp": exp_path,
+                "exp_type": gsm.columns.loc["VALUE", "description"],
+                "platform": gse.gpls[platform_id].metadata["title"][0],
+                "platform_id": platform_id,
+                "species": gsm.metadata["organism_ch1"][0],
+            }
+
+            self.run_process("upload-microarray-expression", process_inputs)
+            ma_data = Data.filter(entity__name=exp_path)[-1]
+            ma_data.entity.name = name
+            unmapped_data.append(ma_data.id)
+
+        names = [gsm for gsm in gse.gsms.keys()]
+        info_df = {"Accession": names, "SampleName": names}
+
+        mapping_inputs = {"expressions": unmapped_data}
+        if inputs.advanced.mapping_file:
+            mapping_file = inputs.advanced.mapping_file.import_file(
+                imported_format="compressed"
+            )
+
+            stem = Path(mapping_file).stem
+            supported_extensions = (".tab", ".tsv", ".txt")
+            if not stem.endswith(supported_extensions):
+                self.error(
+                    "Mapping file has unsupported file name extension. "
+                    f"The supported extensions are {supported_extensions}."
+                )
+
+            mapping_inputs["mapping_file"] = mapping_file
+            if inputs.advanced.source:
+                mapping_inputs["source"] = inputs.advanced.source
+            else:
+                self.error(
+                    "Custom probe id mapping file was provided but no source was selected."
+                )
+            if inputs.advanced.build:
+                mapping_inputs["build"] = inputs.advanced.build
+            else:
+                self.error(
+                    "Custom probe id mapping file was provided but genome build was not defined."
+                )
+
+        self.run_process("map-microarray-probes", mapping_inputs)
+        return pd.DataFrame(info_df).set_index("Accession", drop=False)
+
     def run(self, inputs, outputs):
         """Run the analysis."""
 
@@ -303,6 +397,7 @@ class GeoImport(Process):
         supported = set(
             [
                 "Expression profiling by high throughput sequencing",
+                "Expression profiling by array",
             ]
         )
         gse_type = gse.get_type() if type(gse.get_type()) is list else [gse.get_type()]
@@ -326,6 +421,9 @@ class GeoImport(Process):
             series_type = series.get_type()
             if series_type == "Expression profiling by high throughput sequencing":
                 run_info = self.upload_rna_gse(inputs, series)
+                metadata_tables[series.name] = create_metadata(series, run_info)
+            elif series_type == "Expression profiling by array":
+                run_info = self.upload_ma_gse(inputs, series)
                 metadata_tables[series.name] = create_metadata(series, run_info)
             else:
                 self.warning(